% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHierarchicalCluster.R
\name{buildHierarchicalCluster}
\alias{buildHierarchicalCluster}
\title{Hello, world!}
\usage{
buildHierarchicalCluster(
  dataset,
  distanceMethod = "euclidean",
  linkageMethod = "complete"
)
}
\arguments{
\item{dataset}{An input}

\item{distanceMethod}{An input}

\item{linkageMethod}{An input}
}
\description{
Hello, world!
}
