% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterAssignment.R
\name{getClusterAssignment}
\alias{getClusterAssignment}
\title{Hello, world!}
\usage{
getClusterAssignment(clusterHierachy, nCluster)
}
\arguments{
\item{clusterHierachy}{An input}

\item{nCluster}{An input}
}
\description{
Hello, world!
}
