% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistancePlot.R
\name{getDistancePlot}
\alias{getDistancePlot}
\title{Generate a Plotly Visualization of a Distance Matrix}
\usage{
getDistancePlot(distanceMatrix, clusterAssignment)
}
\arguments{
\item{distanceMatrix}{Observations summarized in a distance matrix}

\item{clusterAssignment}{Cluster assignments matching the observations}
}
\description{
Generate a Plotly Visualization of a Distance Matrix
}
