% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHierarchicalCluster.R
\name{buildHierarchicalCluster}
\alias{buildHierarchicalCluster}
\title{Perform Hierarchical Clustering on Data}
\usage{
buildHierarchicalCluster(
  dataset,
  distanceMethod = "euclidean",
  linkageMethod = "complete"
)
}
\arguments{
\item{dataset}{An input dataset}

\item{distanceMethod}{A dsitane method applicable for dist}

\item{linkageMethod}{A linkage method applicable for hclust}
}
\description{
Perform Hierarchical Clustering on Data
}
