% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterAssignment.R
\name{getClusterAssignment}
\alias{getClusterAssignment}
\title{Assign Hard Clusters for Cluster Tree}
\usage{
getClusterAssignment(clusterHierachy, nCluster)
}
\arguments{
\item{clusterHierachy}{A hierarchical cluster}

\item{nCluster}{Number of clusters requested}
}
\description{
Assign Hard Clusters for Cluster Tree
}
