% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistancePlot.R
\name{getDistancePlot}
\alias{getDistancePlot}
\title{Hello, world!}
\usage{
getDistancePlot(distanceMatrix, clusterHierachy, clusterAssignment)
}
\arguments{
\item{distanceMatrix}{An input}

\item{clusterHierachy}{An input}

\item{clusterAssignment}{An input}
}
\description{
Hello, world!
}
